# docs/openapi/TEST.yml
openapi: 3.0.0
info:
  title: TEST API
  version: 1.0.0
tags:
  - name: test
    description: ...
  - name: calc
    description: 工数計算関連
paths:
  /TEST/calc:
    post:
      summary: 工数計算
      tags:
        - calc
      operationId: calcTestApplication
      description: ...
      x-eov-operation-id: calcTestApplication
      x-eov-operation-handler: calcApi
      requestBody:
        description: ...
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalcTestApplicationRequest"
      responses:
        "200":
          description: ...
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalcTestApplication200Response"

  # /TEST/test:
  #   post:
  #     tags:
  #       - test
  #     summary: テスト
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/User'
  #     responses:
  #       '201':
  #         description: テスト
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/User'

  /TEST/ping:
    get:
      tags:
        - test
      summary: 接続確認用
      operationId: ping
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: pong

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    CalcTestApplicationRequest:
      description: 工数計算リクエスト
      type: object
      properties:
        totalFP:
          type: number
        manMonth:
          type: number
    CalcTestApplication200Response:
      description: 工数計算レスポンス
      type: object
      properties:
        totalFP:
          type: number
        manMonth:
          type: number