# src/api/src/models/calcTestApplicationRequest.ts
/**
 * 見積算出システム API
 * ...
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
* 工数計算リクエスト
*/
export class CalcTestApplicationRequest {
    'totalFP'?: number;
    'manMonth'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "totalFP",
            "baseName": "totalFP",
            "type": "number"
        },
        {
            "name": "manMonth",
            "baseName": "manMonth",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CalcTestApplicationRequest.attributeTypeMap;
    }
}